+++ READMEe.md
@@ -1,0 +1,1 @@
+# games-rating
