+++ appModules/rating/config.js
@@ -1,8 +1,21 @@
-const WEIGHT = "./dataset/rating.json"
+
 const PATH_TO_RATING_FILE = "./dataset/rating.json";
+function createRating(obj, weight) {
+    return (
+      obj.gameplay * weight.gameplay +
+      obj.design * weight.design +
+      obj.idea * weight.idea
+    );
+  }
+  function updateRating(array, id, rating) {
+    const index = array.findIndex((item) => item.id === id);
+    array[index].rating += rating;
+    return array;
+  }
 
 
 module.exports ={
     PATH_TO_RATING_FILE,
-    WEIGHT
+    updateRating,
+    createRating
 }
