+++ appModules/rating/index.js
@@ -1,10 +1,7 @@
 const config = require("./config");
 const makeRatingFile = require("./rating-file");
-const {createRating,updateRating} = require("./calculations")
 
 module.exports ={
     config,
-    makeRatingFile,
-    createRating, 
-    updateRating
+    makeRatingFile
 }
