+++ controllers/vote.js
@@ -1,6 +1,6 @@
 const fs = require("fs").promises;
 const parseBody = require("../appModules/http-utils/parse-body");
-const { config, createRating, updateRating } = require("../appModules/rating");
+const { config, createRating, updateRating } = require("../appModules/rating/config");
 
 async function voteRouteController(req, res) {
   if (req.method !== "POST") {
