+++ dataset/rating.json
@@ -1,1 +1,1 @@
-[{"id":5,"title":"G.U.N.N.E.R.","image":"https://code.s3.yandex.net/teens/pindie-games/gunner/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/gunner/game/index.html","description":"Продержись как можно дольше, отбиваясь от врагов, которых будет становиться всё больше.","rating":null},{"id":3,"title":"Defence of Crystal","image":"https://code.s3.yandex.net/teens/pindie-games/defence-of-crystal/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/defence-of-crystal/game/index.html","description":"Защищай магический кристалл от монстров и уничтожай кладбища, чтобы спасти Землю, которую поглотил мрак.\n","rating":0},{"id":1,"title":"Crystal Keeper","image":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/game/index.html","description":"Управляй боевым дроном, чтобы любой ценой защитить кристалл от враждебных космо-слизней.","rating":0},{"id":4,"title":"Go Away","image":"https://code.s3.yandex.net/teens/pindie-games/go-away/cover.jpg","link":"https://code.s3.yandex.net/teens/pindie-games/go-away/game/index.html","description":"Управляй автомобилем, избегая аварий и перепрыгивая препятствия на пути к следующему уровню.\n","rating":null},{"id":9,"title":"Deep Under The Castle","image":"https://code.s3.yandex.net/teens/pindie-games/Deep-Under-The-Castle/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/Deep-Under-The-Castle/index.html","description":"Прыгай по платформам и собирай ключи, чтобы выбраться из проклятого замка.\n","rating":0},{"id":6,"title":"Space Terror","image":"https://code.s3.yandex.net/teens/pindie-games/space-terror/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/space-terror/game/index.html","description":"Лети как можно дальше в космическом пространстве, уничтожая всё на своём пути.","rating":0},{"id":2,"title":"Dangeons'n'Caves. Prologue","image":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/game/index.html","description":"Безымянный герой исследует пещеры и подземелья, чтобы найти больше информации о себе.","rating":0},{"id":7,"title":"Square Slayer","image":"https://code.s3.yandex.net/teens/pindie-games/square-slayer/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/square-slayer/game/index.html","description":"Уворачивайся и отстреливайся от озлобленных квадратов и собирай жизни для перехода на следующий уровень.","rating":0}]
+[{"id":5,"title":"G.U.N.N.E.R.","image":"https://code.s3.yandex.net/teens/pindie-games/gunner/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/gunner/game/index.html","description":"Продержись как можно дольше, отбиваясь от врагов, которых будет становиться всё больше.","rating":0},{"id":3,"title":"Defence of Crystal","image":"https://code.s3.yandex.net/teens/pindie-games/defence-of-crystal/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/defence-of-crystal/game/index.html","description":"Защищай магический кристалл от монстров и уничтожай кладбища, чтобы спасти Землю, которую поглотил мрак.\n","rating":0},{"id":1,"title":"Crystal Keeper","image":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/game/index.html","description":"Управляй боевым дроном, чтобы любой ценой защитить кристалл от враждебных космо-слизней.","rating":0},{"id":4,"title":"Go Away","image":"https://code.s3.yandex.net/teens/pindie-games/go-away/cover.jpg","link":"https://code.s3.yandex.net/teens/pindie-games/go-away/game/index.html","description":"Управляй автомобилем, избегая аварий и перепрыгивая препятствия на пути к следующему уровню.\n","rating":0},{"id":9,"title":"Deep Under The Castle","image":"https://code.s3.yandex.net/teens/pindie-games/Deep-Under-The-Castle/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/Deep-Under-The-Castle/index.html","description":"Прыгай по платформам и собирай ключи, чтобы выбраться из проклятого замка.\n","rating":0},{"id":6,"title":"Space Terror","image":"https://code.s3.yandex.net/teens/pindie-games/space-terror/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/space-terror/game/index.html","description":"Лети как можно дальше в космическом пространстве, уничтожая всё на своём пути.","rating":0},{"id":2,"title":"Dangeons'n'Caves. Prologue","image":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/game/index.html","description":"Безымянный герой исследует пещеры и подземелья, чтобы найти больше информации о себе.","rating":0},{"id":7,"title":"Square Slayer","image":"https://code.s3.yandex.net/teens/pindie-games/square-slayer/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/square-slayer/game/index.html","description":"Уворачивайся и отстреливайся от озлобленных квадратов и собирай жизни для перехода на следующий уровень.","rating":0}]
